Add pamflet.arrow configuration option for next page icon.
This commit adds a new template.properties configuration option
"pamflet.arrow" which allows the user to set the string used for
the next page icon.  The string (e.g. ">") is used as the next page
icon on the right side of the page and is flipped horizontally to create
the previous page icon (e.g. "<").