Significant refactor of localization structure.
This makes `Content` and `Page` more responsible for knowing how to
generate links, so that the news code can do it in a different way.

Once incidental change here is that css and other custom assets in for
a default language are copied into the language-specific directory for
other languages. This is simpler than treating them as different from
language-specific assets, but there is a downside: the duplicated
assets must be loaded when switching languages, and there is a
potential for a name collision.