Removing the gray background hover on desktop css.
There's a minor layout bug where this background is active below the
content as well when the content is short.

Not sure exactly how to fix that without making the layout more
complicated, but also, the more I look at this behavior the more I
think the background hover gray is just ugly and dated, so, buh bye!