Include links to other stories below all stories.
This makes the navigation of Pamflet News more similar to the original
Pamflet, where the top level navigation is accessible at the bottom of
each page.

Since the the structure and representation of this navigation depends
on the content itself, I decided to move it into the content
representation types.

This goes against an earlier plan to keep HTML out of these types as
much as possible, to consider them to be layout-independent data that
is later rendered into HTML by `Printer`, and could potentially be
rendered into some other format. Too bad for that plan, I guess.